requirements.txt
matplotlib<3.7.1
mysql_connector_repackaged==0.3.1
pandas==1.5.3
seaborn==0.12.2
streamlit==1.12.0
pillow==9.2.0
numpy<1.24.3

